+++ src/Column/DataColumn.php
@@ -36,7 +36,7 @@ use function sprintf;
  * calculation, while the actual cell content is a {@see format|formatted} version of that value which may contain HTML
  * markup.
  */
-final class DataColumn implements ColumnInterface, ColumnRendererInterface
+final class DataColumn implements ColumnInterface
 {
     /** @psalm-var string[] */
     private array $filterTypes = [
@@ -140,7 +140,7 @@ final class DataColumn implements ColumnInterface, ColumnRendererInterface
 
     public function getFilterType(): string
     {
-        return $this->filterType;
+        return $this->filterTypes[$this->filterType];
     }
 
     public function getFilterInputAttributes(): array
@@ -176,178 +176,8 @@ final class DataColumn implements ColumnInterface, ColumnRendererInterface
         return $this->visible;
     }
 
-    private function renderFilterInput(self $column, GlobalContext $context): string
+    public function getRenderer(): string
     {
-        $filterInputAttributes = $column->getFilterInputAttributes();
-        $filterInputTag = Input::tag();
-
-        if (!array_key_exists('name', $filterInputAttributes)) {
-            $filterInputTag = $filterInputTag->name(
-                Attribute::getInputName(
-                    (string)($column->getFilterModelName() ?? $context->getFilterModelName()),
-                    $column->getFilterProperty() ?? ''
-                ),
-            );
-        }
-
-        if (!array_key_exists('value', $filterInputAttributes) && $column->getFilterValueDefault() !== '') {
-            $filterInputTag = $filterInputTag->value($column->getFilterValueDefault());
-        }
-
-        return $filterInputTag
-            ->addAttributes($filterInputAttributes)
-            ->type($this->filterTypes[$this->filterType])
-            ->render();
-    }
-
-    private function renderFilterSelect(self $column, GlobalContext $context): string
-    {
-        $filterInputAttributes = $column->getFilterInputAttributes();
-        $filterSelectTag = Select::tag();
-
-        if (!array_key_exists('name', $filterInputAttributes)) {
-            $filterSelectTag = $filterSelectTag->name(
-                Attribute::getInputName(
-                    (string)($column->getFilterModelName() ?? $context->getFilterModelName()),
-                    $column->getFilterProperty() ?? ''
-                ),
-            );
-        }
-
-        if ($column->getFilterValueDefault() !== null) {
-            $filterSelectTag = $filterSelectTag->value($column->getFilterValueDefault());
-        }
-
-        return $filterSelectTag
-            ->addAttributes($filterInputAttributes)
-            ->optionsData($column->getFilterInputSelectItems())
-            ->prompt($column->getFilterInputSelectPrompt())
-            ->render();
-    }
-
-    public function getRenderer(): self
-    {
-        return $this;
-    }
-
-    public function renderColumn(ColumnInterface $column, Cell $cell, GlobalContext $context): Cell
-    {
-        $this->checkColumn($column);
-        return $cell->addAttributes($column->getColumnAttributes());
-    }
-
-    public function renderHeader(ColumnInterface $column, Cell $cell, GlobalContext $context): Cell
-    {
-        $this->checkColumn($column);
-
-        $label = $this->getHeader() ?? ($this->property === null ? '' : ucfirst($this->property));
-
-        if ($column->getProperty() !== null && $column->isWithSorting()) {
-            $linkSorter = $this->renderLinkSorter($context, $column->getProperty(), $label);
-            if (!empty($linkSorter)) {
-                return $cell->content($linkSorter)->encode(false);
-            }
-        }
-
-        return $cell
-            ->addAttributes($column->getHeaderAttributes())
-            ->content($label);
-    }
-
-    public function renderFilter(ColumnInterface $column, Cell $cell, GlobalContext $context): ?Cell
-    {
-        $this->checkColumn($column);
-
-        if ($column->getFilter() !== null) {
-            $content = $column->getFilter();
-        } elseif ($column->getFilterProperty() !== null) {
-            $content = match ($column->getFilterType()) {
-                'select' => $this->renderFilterSelect($column, $context),
-                default => $this->renderFilterInput($column, $context),
-            };
-        } else {
-            return null;
-        }
-
-        return $cell
-            ->content($content)
-            ->addAttributes($column->getFilterAttributes())
-            ->encode(false);
-    }
-
-    public function renderBody(ColumnInterface $column, Cell $cell, DataContext $context): Cell
-    {
-        $this->checkColumn($column);
-
-        $contentSource = $column->getContent();
-
-        if ($contentSource !== null) {
-            $content = (string)(is_callable($contentSource) ? $contentSource($context) : $contentSource);
-        } elseif ($this->property !== null) {
-            $content = (string)ArrayHelper::getValue($context->getData(), $this->property);
-        } else {
-            $content = '';
-        }
-
-        return $cell
-            ->addAttributes($column->getBodyAttributes())
-            ->content($content);
-    }
-
-    public function renderFooter(ColumnInterface $column, Cell $cell, GlobalContext $context): Cell
-    {
-        $this->checkColumn($column);
-
-        if ($column->getFooter() !== null) {
-            $cell = $cell->content($column->getFooter());
-        }
-
-        return $cell;
-    }
-
-    private function renderLinkSorter(GlobalContext $context, string $property, string $label): string
-    {
-        $dataReader = $context->getDataReader();
-        if (!$dataReader instanceof PaginatorInterface) {
-            return '';
-        }
-
-        $sort = $dataReader->getSort();
-        if ($sort === null) {
-            return '';
-        }
-
-        $linkSorter = $dataReader instanceof OffsetPaginator
-            ? LinkSorter::widget()->currentPage($dataReader->getCurrentPage())
-            : LinkSorter::widget();
-
-        return $linkSorter
-            ->attribute($property)
-            ->attributes($sort->getCriteria())
-            ->directions($sort->getOrder())
-            ->iconAscClass('bi bi-sort-alpha-up')
-            ->iconDescClass('bi bi-sort-alpha-down')
-            ->label($label)
-            ->linkAttributes($context->getSortLinkAttributes())
-            ->pageSize($dataReader->getPageSize())
-            ->urlArguments($context->getUrlArguments())
-            ->urlQueryParameters($context->getUrlQueryParameters())
-            ->render();
-    }
-
-    /**
-     * @psalm-assert self $column
-     */
-    private function checkColumn(ColumnInterface $column): void
-    {
-        if (!$column instanceof self) {
-            throw new InvalidArgumentException(
-                sprintf(
-                    'Expected "%s", but "%s" given.',
-                    self::class,
-                    $column::class
-                )
-            );
-        }
+        return DataColumnRenderer::class;
     }
 }
